Prevent CSRF attack on user profile deletion

+++ modules/unreguser/unreguser.php
@@ -43,12 +43,15 @@ if (!isset($doit) or $doit != "yes") {
 		draw($tool_content,1);
 		exit;
 	} else {
+		if(isset($_SESSION['token'])) {
+			$CSRFToken = $_SESSION['token'];
+		}
 		$q = db_query ("SELECT code FROM cours, cours_user WHERE cours.cours_id = cours_user.cours_id AND user_id = '$uid' LIMIT 1") ;
 		if (mysql_num_rows($q) == 0) {
 			$tool_content .=  "<p><b>$langConfirm</b></p>";
 			$tool_content .=  "<ul class=\"listBullet\">";
 			$tool_content .=  "<li>$langYes: ";
-			$tool_content .=  "<a href='$_SERVER[PHP_SELF]?u=$uid&doit=yes'>$langDelete</a>";
+			$tool_content .=  "<a href='$_SERVER[PHP_SELF]?u=$uid&doit=yes&CSRFToken=$CSRFToken'>$langDelete</a>";
 			$tool_content .=  "</li>";
 			$tool_content .=  "<li>$langNo: <a href='../profile/profile.php'>$langBack</a>";
 			$tool_content .=  "</li></ul>";
@@ -62,15 +65,17 @@ if (!isset($doit) or $doit != "yes") {
 	}  //endif is admin
 } else {
 	if (isset($uid)) {
-		$tool_content .=  "<table width=99%><tbody>";
-		$tool_content .=  "<tr>";
-		$tool_content .=  "<td class=\"success\">";
-		db_query("DELETE from user WHERE user_id = '$uid'");
-		if (mysql_affected_rows() > 0) {
-			$tool_content .=  "<p><b>$langDelSuccess</b></p>";
-			$tool_content .=  "<p>$langThanks</p>";
-			$tool_content .=  "<br><a href='../../index.php?logout=yes'>$langLogout</a>";
-			unset($_SESSION['uid']);
+		if(isset($CSRFToken) and $CSRFToken == $_SESSION['token']) {
+			$tool_content .=  "<table width=99%><tbody>";
+			$tool_content .=  "<tr>";
+			$tool_content .=  "<td class=\"success\">";
+			db_query("DELETE from user WHERE user_id = '$uid'");
+			if (mysql_affected_rows() > 0) {
+				$tool_content .= "<p><b>$langDelSuccess</b></p>";
+				$tool_content .= "<p>$langThanks</p>";
+				$tool_content .= "<br><a href='../../index.php?logout=yes'>$langLogout</a>";
+				unset($_SESSION['uid']);
+			}
 		} else {
 			$tool_content .=  "<p>$langError</p>";
 			$tool_content .=  "<p><a href='../profile/profile.php'>$langBack</a></p><br>";
